requirements.txt
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.6.0
beautifulsoup4==4.13.4
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.3
dataclasses-json==0.6.7
exceptiongroup==1.3.0
fastmcp==2.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
jmapc==0.2.23
markdown-it-py==3.0.0
marshmallow==3.26.1
mcp==1.9.3
mdurl==0.1.2
mypy_extensions==1.1.0
openapi-pydantic==0.5.1
packaging==25.0
pip==25.1.1
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pydantic-settings==2.9.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
requests==2.32.3
rich==14.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sse-starlette==2.3.6
sseclient==0.0.27
starlette==0.47.0
typer==0.16.0
typing_extensions==4.14.0
typing-inspect==0.9.0
typing-inspection==0.4.1
urllib3==2.4.0
uvicorn==0.34.3
Flask==3.0.3
openai>=1.0.0
